requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
alembic==1.10.3
python-dotenv==1.0.0
pydantic==1.10.7
celery==5.2.7
redis==4.5.3
requests==2.28.2

sqlalchemy==2.0.10
psycopg2-binary==2.9.6
passlib==1.7.4
pyjwt==2.6.0
## The following requirements were added by pip freeze:
amqp==5.3.1
anyio==4.5.2
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
exceptiongroup==1.2.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
kombu==5.5.2
Mako==1.3.10
MarkupSafe==2.1.5
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pdf2image==1.17.0
pillow==10.4.0
prompt_toolkit==3.0.50
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tesseract==0.1.3
typing_extensions
tzdata==2025.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.2
uvicorn==0.22.0
python-dotenv==1.0.0